test wmo course working with ERA5 data and simulation 2007-06-22

+++ scripts/make_degrib.bash
@@ -67,17 +67,23 @@ cp -f /usr/lib64/libjpeg.so* ${HOME}/local/lib64
 if [ ! -s ${BNDDIR}/gfs.t${YYYYMMDDHHi:8:2}z.pgrb2.0p25.f000.${YYYYMMDDHHi}.grib2 ]
 then
    if [ ! -s ${GCCCIS}/${YYYYMMDDHHi:0:4}/${YYYYMMDDHHi}/gfs.t${YYYYMMDDHHi:8:2}z.pgrb2.0p25.f000.${YYYYMMDDHHi}.grib2 ]
-   then
-      echo -e "${RED}==>${NC}Condicao de contorno inexistente !"
-      echo -e "${RED}==>${NC}Check ${BNDDIR} or." 
-      echo -e "${RED}==>${NC}Check ${GCCCIS}"
-      exit 1            
+   then 
+      if [ ! -s /mnt/beegfs/guilherme.mendonca/MPAS-BR/met_data/ERA5/DATA/'era5_pl_2007-06-22 00:00:00_0000.grib' ]
+      then
+         echo -e "${RED}==>${NC}Condicao de contorno inexistente !"
+         echo -e "${RED}==>${NC}Check ${BNDDIR} or." 
+         echo -e "${RED}==>${NC}Check ${GCCCIS}"
+         exit 1
+      else
+         BNDDIR=/mnt/beegfs/guilherme.mendonca/MPAS-BR/met_data/ERA5/DATA/	      
+         filename=
+      fi  
    else
       BNDDIR=${GCCCIS}/${YYYYMMDDHHi:0:4}/${YYYYMMDDHHi}
    fi    
 fi
 
-files_needed=("${DATAIN}/fixed/${MESH}.static.nc" "${DATAIN}/fixed/Vtable.${EXP}" "${EXECS}/ungrib.exe" "${BNDDIR}/gfs.t${YYYYMMDDHHi:8:2}z.pgrb2.0p25.f000.${YYYYMMDDHHi}.grib2")
+files_needed=("${DATAIN}/fixed/${MESH}.static.nc" "${DATAIN}/fixed/Vtable.${EXP}" "${EXECS}/ungrib.exe" "${BNDDIR}/gfs.t${YYYYMMDDHHi:8:2}z.pgrb2.0p25.f000.${YYYYMMDDHHi}.grib2 ")
 for file in "${files_needed[@]}"
 do
   if [ ! -s "${file}" ]
